requirements.txt
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
Django==2.1.5
django-extensions==2.1.5
django-markdown-deux==1.0.5
django-pagedown==1.0.6
idna==2.8
kiwisolver==1.0.1
lti==0.9.4
lxml==4.3.0
markdown2==2.3.7
matplotlib==3.0.2
numpy==1.16.1
oauthlib==3.0.0
Pillow==5.4.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
urllib3==1.24.1
